Added Climate state and changed timestamps to Long to fix exponentiation problem

+++ src/main/java/com/tesla/springboot/evapi/entities/ChargeState.java
@@ -177,5 +177,5 @@ public class ChargeState extends AbstractEntity{
 
     @Column(name="time_stamp")
     @JsonProperty("timestamp")
-    Double timestamp;
+    Long timestamp;
 }
